console: Refactor managed gateway by new UI
