+++ pkg/webui/lib/shared-messages.js
@@ -138,7 +138,6 @@ export default defineMessages({
   connected: 'Connected',
   connecting: 'Connecting',
   connectionIssues: 'Connection issues',
-  connectionProfiles: 'Connection profiles',
   connectionSettings: 'Connection settings',
   contactFieldPlaceholder: 'Type or choose a collaborator',
   contactInformation: 'Contact information',
@@ -581,4 +580,5 @@ export default defineMessages({
   webhookId: 'Webhook ID',
   webhookUpdated: 'Webhook updated',
   webhooks: 'Webhooks',
+  wifiProfiles: 'WiFi profiles',
 })
